ciphergrid_backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
kafka-python==2.0.2



